zebra: ensure a new FDB entry is permanent
In ab4d663d507a, the state of a new FDB entry was set to NUD_NOARP
only. However, the kernel require a state to be either NUD_PERMANENT or
NUD_REACHABLE. This is enforced by iproute2. Recent versions default to
NUD_REACHABLE, but in our case, NUD_PERMANENT makes more sense.

Without this change, we get:

    ZEBRA: netlink-cmd (NS 0) error: Invalid argument, type=RTM_NEWNEIGH(28), seq=31, pid=0
    vxlan: RTM_NEWNEIGH with invalid state 0x40